Add a FrameReader to extract HDLC frames from a ByteStream
This function will take as any struct that implements READ plus a prefix
and returns a vec of hdlc frames
+ the rest of the data.

This will allow to read data e.g. from a device regularly and get the frames out, the rest of the last call then then be used as a prefix for the next one.