Modified to allow extra bytes to be escaped
This was needed as sensiron SPS30 sensors use a modified version of HDLC
that escapes extra bytes. This is done by using a HashMap as the list of
translation characters

Please note this has broken decode_slice. Hopefully I will fix that
eventually. But for now this is just a weird version for a special
use-case